requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
pypdf==5.4.0
pydantic==2.11.3
python-multipart==0.0.20
